ramfs: Extract common ramfs_create()
