+++ src/fs/driver/ramfs/ramfs.h
@@ -57,6 +57,7 @@ extern int ramfs_fill_sb(struct super_block *sb, const char *source);
 extern int ramfs_format(struct block_dev *bdev, void *priv);
 extern int ramfs_delete(struct inode *node);
 extern int ramfs_truncate(struct inode *node, off_t length);
+extern int ramfs_create(struct inode *i_new, struct inode *i_dir, int mode);
 
 extern struct ramfs_file_info *ramfs_file_alloc(struct inode *node);
