+++ src/fs/driver/ramfs/ramfs_dvfs.c
@@ -66,27 +66,6 @@ static int ramfs_iterate(struct inode *next, char *name, struct inode *parent, s
 	return -1;
 }
 
-static int ramfs_create(struct inode *i_new, struct inode *i_dir, int mode) {
-	struct ramfs_file_info *fi;
-
-	if (S_ISREG(i_new->i_mode)) {
-		assert(i_new);
-		assert(i_new->i_dentry);
-		assert(i_new->i_dentry->name);
-
-		fi = ramfs_file_alloc(i_new);
-		if (NULL == fi) {
-			return -ENOMEM;
-		}
-		fi->mode = i_new->i_mode;
-		strncpy(fi->name, inode_name(i_new), sizeof(fi->name) - 1);
-
-		i_new->i_no = fi->index;
-	}
-
-	return 0;
-}
-
 static struct inode *ramfs_ilookup(char const *name, struct inode const *dir) {
 	struct inode *node;
 	struct super_block *sb;
